Remove AccountReset::NotifyUserOfRequestCancellation
- Move logic into AccountResetRequest#cancel!
