+++ app/models/account_reset_request.rb
@@ -25,7 +25,8 @@ class AccountResetRequest < ApplicationRecord
 
   def cancel!(now: Time.zone.now)
     update(cancelled_at: now)
-    AccountReset::NotifyUserOfRequestCancellation.new(user).call
+    notify_user_via_email_of_account_reset_cancellation
+    notify_user_via_phone_of_account_reset_cancellation
   end
 
   def granted_token_valid?
@@ -36,5 +37,24 @@ class AccountResetRequest < ApplicationRecord
     granted_at.present? &&
       ((Time.zone.now - granted_at) >
        IdentityConfig.store.account_reset_token_valid_for_days.days)
+  end
+
+  private
+
+  def notify_user_via_email_of_account_reset_cancellation
+    user.confirmed_email_addresses.each do |email_address|
+      UserMailer.with(user: user, email_address: email_address).account_reset_cancel.
+        deliver_now_or_later
+    end
+  end
+
+  def notify_user_via_phone_of_account_reset_cancellation
+    MfaContext.new(user).phone_configurations.each do |phone_configuration|
+      phone = phone_configuration.phone
+      Telephony.send_account_reset_cancellation_notice(
+        to: phone,
+        country_code: Phonelib.parse(phone).country,
+      )
+    end
   end
 end
