+++ spec/models/account_reset_request_spec.rb
@@ -135,14 +135,29 @@ RSpec.describe AccountResetRequest do
       expect(other_request.reload.cancelled_at).to be_nil
     end
 
-    it "sends an email to the user's confirmed email addresses" do
-      notify_user_of_cancellation = instance_double(AccountReset::NotifyUserOfRequestCancellation)
-      expect(AccountReset::NotifyUserOfRequestCancellation).to receive(:new).
-        with(user).
-        and_return(notify_user_of_cancellation)
-      expect(notify_user_of_cancellation).to receive(:call)
+    it "sends an email to the user's confirmed email addresses and phone numbers" do
+      email_address1 = user.email_addresses.first
+      email_address2 = create(:email_address, user: user)
+
+      phone_config1 = create(:phone_configuration, user: user)
+      phone_config2 = create(:phone_configuration, user: user)
+
+      expect(Telephony).to receive(:send_account_reset_cancellation_notice).
+        with(to: phone_config1.phone, country_code: 'US')
+      expect(Telephony).to receive(:send_account_reset_cancellation_notice).
+        with(to: phone_config2.phone, country_code: 'US')
 
       cancel!
+
+      expect_delivered_email_count(2)
+      expect_delivered_email(
+        to: [email_address1.email],
+        subject: t('user_mailer.account_reset_cancel.subject'),
+      )
+      expect_delivered_email(
+        to: [email_address2.email],
+        subject: t('user_mailer.account_reset_cancel.subject'),
+      )
     end
   end
 end
